+++ include/hep/mc-mpi.hpp
@@ -24,6 +24,7 @@
 #include "hep/mc/mpi_multi_channel.hpp"
 #include "hep/mc/mpi_multi_channel_callback.hpp"
 #include "hep/mc/mpi_plain.hpp"
+#include "hep/mc/mpi_plain_callback.hpp"
 #include "hep/mc/mpi_vegas.hpp"
 #include "hep/mc/mpi_vegas_callback.hpp"